pyproject.toml
[tool.poetry]
name = "micromarket-"
version = "0.1.0"
description = ""
authors = ["Anakion(Anatoliy)"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
